move: 유저 프로필 수정 DTO 디렉토리 이동 (#CHZZ-93)
