+++ src/main/java/org/chzz/market/domain/user/dto/request/UpdateUserProfileRequest.java
@@ -1,4 +1,4 @@
-package org.chzz.market.domain.user.dto;
+package org.chzz.market.domain.user.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
