+++ src/main/java/org/chzz/market/domain/user/service/UserService.java
@@ -8,8 +8,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.chzz.market.domain.auction.dto.response.AuctionParticipationResponse;
 import org.chzz.market.domain.auction.repository.AuctionRepository;
-import org.chzz.market.domain.user.dto.UpdateProfileResponse;
-import org.chzz.market.domain.user.dto.UpdateUserProfileRequest;
+import org.chzz.market.domain.user.dto.response.UpdateProfileResponse;
+import org.chzz.market.domain.user.dto.request.UpdateUserProfileRequest;
 import org.chzz.market.domain.user.dto.request.UserCreateRequest;
 import org.chzz.market.domain.user.dto.response.NicknameAvailabilityResponse;
 import org.chzz.market.domain.user.dto.response.ParticipationCountsResponse;
